using HashcashPaymaster
This Paymaster shows how on-chain contract requries some work from the
off-chain provider.
In this case, it requires a has  over the sender's nonce and address
that is hard to calculate - much like block mining, to avoid spamming
the API.
difficulty=15 (as defined in `depl.js` deployment script) requires few
seconds to calculate on a desktop machine.