Add 'source-map-explorer' dependency and 'analyze' script
We have not paid attention to the size increase of the bundle GSN causes
for its users. It has become an issue now that Web3.js is no longer
widespread and GSN is often the only reason it is imported.

Executing 'analyze' creates a report on the size of the GSN library.